@@ -1,73 +1,44 @@
 {
-    "name": "fatfelines",
-    "version": "1.0.0",
-    "license": "MIT",
-    "author": "danydodson",
-    "repository": "github:danydodson/fatfelines",
-    "private": true,
-    "main": "index.js",
-    "scripts": {
-        "start": "nodemon --watch build --exec \"webpack-dev-server --config config/webpack.dev.js\"",
-        "build": "rimraf dist && webpack --config config/webpack.prod.js",
-        "lint": "pretty-quick --staged && ESLINT_USE_FLAT_CONFIG=true npx eslint . -c eslint.config.mjs",
-        "tsc": "tsc"
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "pretty-quick --staged && eslint . --ext .js,.jsx"
-        }
-    },
-    "dependencies": {
-        "@headlessui/react": "^1.7.13",
-        "axios": "^1.3.4",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-lazy-load-image-component": "^1.5.6",
-        "react-router-dom": "^6.9.0",
-        "react-transition-group": "^4.4.5",
-        "tailwind-scrollbar-hide": "^1.1.7"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.21.3",
-        "@babel/eslint-parser": "^7.21.3",
-        "@babel/plugin-proposal-class-properties": "^7.18.6",
-        "@babel/plugin-proposal-decorators": "^7.21.0",
-        "@babel/plugin-proposal-private-methods": "^7.18.6",
-        "@babel/plugin-proposal-private-property-in-object": "^7.21.0",
-        "@babel/preset-env": "^7.20.2",
-        "@babel/preset-react": "^7.18.6",
-        "@types/node": "^18.15.3",
-        "@types/react": "^18.0.28",
-        "@types/react-dom": "^18.0.11",
-        "autoprefixer": "^10.4.14",
-        "babel-loader": "^9.1.2",
-        "css-loader": "^6.7.3",
-        "eslint": "8.36.0",
-        "eslint-config-prettier": "^8.7.0",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-jsx-a11y": "^6.7.1",
-        "eslint-plugin-react": "^7.32.2",
-        "globals": "^13.20.0",
-        "html-webpack-plugin": "^5.5.0",
-        "husky": "^8.0.3",
-        "jest": "^29.5.0",
-        "mini-css-extract-plugin": "^2.7.5",
-        "nodemon": "^2.0.21",
-        "postcss": "^8.4.21",
-        "postcss-cli": "^10.1.0",
-        "postcss-loader": "^7.1.0",
-        "prettier": "^2.8.5",
-        "prettier-plugin-import-sort": "^0.0.7",
-        "pretty-quick": "^3.1.3",
-        "react-app-env": "^1.2.3",
-        "style-loader": "^3.3.2",
-        "tailwind-scrollbar": "^3.0.0",
-        "tailwindcss": "^3.2.7",
-        "typescript": "^5.0.2",
-        "url-loader": "^4.1.1",
-        "webpack": "^5.76.2",
-        "webpack-cli": "^5.0.1",
-        "webpack-dev-server": "^4.13.1",
-        "webpack-merge": "^5.8.0"
-    }
+  "name": "fatcats.com",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
+  "pre-commit": "check",
+  "dependencies": {
+    "@headlessui/react": "^1.7.18",
+    "axios": "^1.6.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.13",
+    "react-lazy-load-image-component": "^1.6.0",
+    "react-router-dom": "^6.22.3",
+    "react-transition-group": "^4.4.5",
+    "tailwind-scrollbar": "^3.1.0",
+    "tailwind-scrollbar-hide": "^1.1.7"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.6",
+    "nodemon": "^3.1.0",
+    "postcss": "^8.4.38",
+    "postcss-cli": "^11.0.0",
+    "postcss-loader": "^8.1.1",
+    "pre-commit": "^1.2.2",
+    "prettier": "^3.2.5",
+    "tailwindcss": "^3.4.3",
+    "prettier-plugin-tailwindcss": "^0.5.12",
+    "vite": "^5.2.0"
+  }
 }
