@@ -1,7 +1,6 @@
-import React from 'react'
 
 const Divider = () => {
-    return <div className='hidden md:block w-[1px] h-8 bg-neutral-400'></div>
+  return <div className='hidden md:block w-[1px] h-8 bg-neutral-400'></div>
 }
 
 export default Divider
