@@ -1,11 +1,11 @@
-import React, { Fragment } from 'react'
+import { Fragment } from 'react'
 
 const Layout = ({ children }) => {
-    return (
-        <Fragment>
-            {children}
-        </Fragment>
-    )
+  return (
+    <Fragment>
+      {children}
+    </Fragment>
+  )
 }
 
 export default Layout
