@@ -1,11 +1,10 @@
-import React from 'react'
 
 const Main = ({ children }) => {
-    return (
-        <div className='px-4 space-y-6 overflow-hidden'>
-            {children}
-        </div>
-    )
+  return (
+    <div className='px-4 space-y-6 overflow-hidden'>
+      {children}
+    </div>
+  )
 }
 
 export default Main
