@@ -1,13 +1,12 @@
-import React from 'react'
 import spinner from '../assets/spinner.gif'
 
 function Spinner() {
-    return (
-        <img src={spinner}
-            className='block w-52 m-auto'
-            alt='Loading...'
-        />
-    )
+  return (
+    <img src={spinner}
+      className='block w-52 m-auto'
+      alt='Loading...'
+    />
+  )
 }
 
 export default Spinner
