@@ -1,26 +1,25 @@
-import React from 'react'
 
 const BarsIcon = () => (
-    <svg id='bars_icon'
-        width='24'
-        height='24'
-        viewBox='0 0 24 24'
-        fill='none'
-        strokeWidth={1.5}
-        stroke='currentColor'
-        className='w-6 h-6'
-        alt=''
-        aria-hidden='true'
-        xmlns='http://www.w3.org/2000/svg'
-        version='1.1'
-    >
-        <title>Bars Icon</title>
-        <path
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            d='M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5'
-        />
-    </svg>
+  <svg id='bars_icon'
+    width='24'
+    height='24'
+    viewBox='0 0 24 24'
+    fill='none'
+    strokeWidth={1.5}
+    stroke='currentColor'
+    className='w-6 h-6'
+    alt=''
+    aria-hidden='true'
+    xmlns='http://www.w3.org/2000/svg'
+    version='1.1'
+  >
+    <title>Bars Icon</title>
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      d='M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5'
+    />
+  </svg>
 )
 
 export default BarsIcon
