@@ -1,33 +1,32 @@
-import React from 'react'
 
 const EllipsisIcon = () => (
-    <svg id='elipsis_icon'
-        width='24'
-        height='24'
-        viewBox='0 0 24 24'
-        fill='none'
-        alt=''
-        aria-hidden='true'
-        xmlns='http://www.w3.org/2000/svg'
-        version='1.1'
-    >
-        <title>Ellipsis Icon</title>
-        <path d='M6.75 12C6.75 12.4142 6.41421 12.75 6 12.75C5.58579 12.75 5.25 12.4142 5.25 12C5.25 11.5858 5.58579 11.25 6 11.25C6.41421 11.25 6.75 11.5858 6.75 12Z'
-            stroke='#0F172A'
-            strokeWidth='1.5'
-            strokeLinecap='round'
-            strokeLinejoin='round' />
-        <path d='M12.75 12C12.75 12.4142 12.4142 12.75 12 12.75C11.5858 12.75 11.25 12.4142 11.25 12C11.25 11.5858 11.5858 11.25 12 11.25C12.4142 11.25 12.75 11.5858 12.75 12Z'
-            stroke='#0F172A'
-            strokeWidth='1.5'
-            strokeLinecap='round'
-            strokeLinejoin='round' />
-        <path d='M18.75 12C18.75 12.4142 18.4142 12.75 18 12.75C17.5858 12.75 17.25 12.4142 17.25 12C17.25 11.5858 17.5858 11.25 18 11.25C18.4142 11.25 18.75 11.5858 18.75 12Z'
-            stroke='#0F172A'
-            strokeWidth='1.5'
-            strokeLinecap='round'
-            strokeLinejoin='round' />
-    </svg>
+  <svg id='elipsis_icon'
+    width='24'
+    height='24'
+    viewBox='0 0 24 24'
+    fill='none'
+    alt=''
+    aria-hidden='true'
+    xmlns='http://www.w3.org/2000/svg'
+    version='1.1'
+  >
+    <title>Ellipsis Icon</title>
+    <path d='M6.75 12C6.75 12.4142 6.41421 12.75 6 12.75C5.58579 12.75 5.25 12.4142 5.25 12C5.25 11.5858 5.58579 11.25 6 11.25C6.41421 11.25 6.75 11.5858 6.75 12Z'
+      stroke='#0F172A'
+      strokeWidth='1.5'
+      strokeLinecap='round'
+      strokeLinejoin='round' />
+    <path d='M12.75 12C12.75 12.4142 12.4142 12.75 12 12.75C11.5858 12.75 11.25 12.4142 11.25 12C11.25 11.5858 11.5858 11.25 12 11.25C12.4142 11.25 12.75 11.5858 12.75 12Z'
+      stroke='#0F172A'
+      strokeWidth='1.5'
+      strokeLinecap='round'
+      strokeLinejoin='round' />
+    <path d='M18.75 12C18.75 12.4142 18.4142 12.75 18 12.75C17.5858 12.75 17.25 12.4142 17.25 12C17.25 11.5858 17.5858 11.25 18 11.25C18.4142 11.25 18.75 11.5858 18.75 12Z'
+      stroke='#0F172A'
+      strokeWidth='1.5'
+      strokeLinecap='round'
+      strokeLinejoin='round' />
+  </svg>
 )
 
 export default EllipsisIcon
