@@ -1,9 +1,9 @@
-export { default as Icon } from './icon'
-export { default as BarsIcon } from './bars'
-export { default as ChatIcon } from './chat'
-export { default as DownIcon } from './down'
-export { default as EllipsisIcon } from './ellipsis'
-export { default as HeartIcon } from './heart'
-export { default as LogoIcon } from './logo'
-export { default as PlusIcon } from './plus'
-export { default as SearchIcon } from './search'
+export { default as Icon } from './Icon'
+export { default as BarsIcon } from './Bars'
+export { default as ChatIcon } from './Chat'
+export { default as DownIcon } from './Down'
+export { default as EllipsisIcon } from './Ellipsis'
+export { default as HeartIcon } from './Heart'
+export { default as LogoIcon } from './Logo'
+export { default as PlusIcon } from './Plus'
+export { default as SearchIcon } from './Search'
