@@ -1,21 +1,20 @@
-import React from 'react'
 
 const LogoIcon = () => (
-    <svg id='logo_icon'
-        width='24'
-        height='24'
-        viewBox='0 0 24 24'
-        fill='#0F172A'
-        className='block h-10 pl-2 w-auto sm:pl-0'
-        xmlns='http://www.w3.org/2000/svg'
-        version='1.1'
-    >
-        <title>Logo Icon</title>
-        <g transform='translate(0.000000,512.000000) scale(0.100000,-0.100000)'>
-            <path d='M139.91,5072.05c-1.83-1.56-9.86-11.11-9.86-11.68c0-0.26,3.44-0.47,7.72-0.47c4.22,0,9.39-0.16,11.47-0.42 c4.59-0.47,4.54-0.52,1.88,6.31c-2.24,5.89-3.81,7.46-7.41,7.46C141.94,5073.25,140.9,5072.94,139.91,5072.05z' />
-            <path d='M107.15,5069.7c-3.34-1.67-3.65-2.4-6.62-16.27c-0.73-3.6-1.46-6.78-1.51-7.09c-0.1-0.37-2.29-2.09-4.9-3.86 c-15.75-10.8-26.34-26.81-30.15-45.53c-1.2-5.95-1.2-21.02,0-26.44c1.88-8.61,5.01-15.91,9.49-22.22c1.3-1.77,2.35-3.49,2.35-3.76 s-3.44-0.63-7.67-0.78c-17.26-0.73-30.82-4.75-40.73-12.05c-3.34-2.45-7.41-7.82-7.41-9.81c0-0.89,2.14-0.94,64.31-0.94 c65.87,0.05,67.12,0.05,76.77,2.03c16.59,3.34,31.5,12.31,40.47,24.2c7.88,10.43,11.42,21.8,11.42,36.3 c-0.05,20.03-6.78,36.25-20.81,50.12c-8.19,8.14-14.45,12.31-24.3,16.32c-9.44,3.81-16.22,5.06-28.16,5.32l-9.39,0.21l-7.46,6.68 c-4.12,3.65-8.24,7.09-9.13,7.61C111.64,5070.9,109.4,5070.9,107.15,5069.7z M170.21,5046.23c6.42-3.18,4.95-12.31-2.19-13.46 c-4.54-0.78-9.13,4.33-7.77,8.61c0.73,2.4,1.46,3.34,3.7,4.69C166.14,5047.38,167.76,5047.43,170.21,5046.23z M152.32,5040.65 c1.46-1.67,1.72-2.4,1.72-4.59c0-4.28-2.92-7.15-7.2-7.15c-2.97,0-6.31,3.02-6.73,6.05c-0.37,2.71,0.89,5.53,3.13,6.99 c1.3,0.83,2.29,1.04,4.49,0.89C150.29,5042.63,150.81,5042.43,152.32,5040.65z' />
-        </g>
-    </svg>
+  <svg id='logo_icon'
+    width='24'
+    height='24'
+    viewBox='0 0 24 24'
+    fill='#0F172A'
+    className='block h-10 pl-2 w-auto sm:pl-0'
+    xmlns='http://www.w3.org/2000/svg'
+    version='1.1'
+  >
+    <title>Logo Icon</title>
+    <g transform='translate(0.000000,512.000000) scale(0.100000,-0.100000)'>
+      <path d='M139.91,5072.05c-1.83-1.56-9.86-11.11-9.86-11.68c0-0.26,3.44-0.47,7.72-0.47c4.22,0,9.39-0.16,11.47-0.42 c4.59-0.47,4.54-0.52,1.88,6.31c-2.24,5.89-3.81,7.46-7.41,7.46C141.94,5073.25,140.9,5072.94,139.91,5072.05z' />
+      <path d='M107.15,5069.7c-3.34-1.67-3.65-2.4-6.62-16.27c-0.73-3.6-1.46-6.78-1.51-7.09c-0.1-0.37-2.29-2.09-4.9-3.86 c-15.75-10.8-26.34-26.81-30.15-45.53c-1.2-5.95-1.2-21.02,0-26.44c1.88-8.61,5.01-15.91,9.49-22.22c1.3-1.77,2.35-3.49,2.35-3.76 s-3.44-0.63-7.67-0.78c-17.26-0.73-30.82-4.75-40.73-12.05c-3.34-2.45-7.41-7.82-7.41-9.81c0-0.89,2.14-0.94,64.31-0.94 c65.87,0.05,67.12,0.05,76.77,2.03c16.59,3.34,31.5,12.31,40.47,24.2c7.88,10.43,11.42,21.8,11.42,36.3 c-0.05,20.03-6.78,36.25-20.81,50.12c-8.19,8.14-14.45,12.31-24.3,16.32c-9.44,3.81-16.22,5.06-28.16,5.32l-9.39,0.21l-7.46,6.68 c-4.12,3.65-8.24,7.09-9.13,7.61C111.64,5070.9,109.4,5070.9,107.15,5069.7z M170.21,5046.23c6.42-3.18,4.95-12.31-2.19-13.46 c-4.54-0.78-9.13,4.33-7.77,8.61c0.73,2.4,1.46,3.34,3.7,4.69C166.14,5047.38,167.76,5047.43,170.21,5046.23z M152.32,5040.65 c1.46-1.67,1.72-2.4,1.72-4.59c0-4.28-2.92-7.15-7.2-7.15c-2.97,0-6.31,3.02-6.73,6.05c-0.37,2.71,0.89,5.53,3.13,6.99 c1.3,0.83,2.29,1.04,4.49,0.89C150.29,5042.63,150.81,5042.43,152.32,5040.65z' />
+    </g>
+  </svg>
 )
 
 export default LogoIcon
