@@ -1,23 +1,22 @@
-import React from 'react'
 
 const SearchIcon = () => (
-    <svg id='search_icon'
-        width='24'
-        height='24'
-        viewBox='0 0 24 24'
-        className='w-6 h-6 text-neutral-500 group-hover:text-neutral-800'
-        xmlns='http://www.w3.org/2000/svg'
-    >
-        <title>Search Icon</title>
-        <path
-            fill='none'
-            stroke='currentColor'
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            strokeWidth='1.5'
-            d='m21 21l-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z'
-        />
-    </svg>
+  <svg id='search_icon'
+    width='24'
+    height='24'
+    viewBox='0 0 24 24'
+    className='w-6 h-6 text-neutral-500 group-hover:text-neutral-800'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <title>Search Icon</title>
+    <path
+      fill='none'
+      stroke='currentColor'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth='1.5'
+      d='m21 21l-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z'
+    />
+  </svg>
 
 )
 
