@@ -5,13 +5,12 @@ import './styles/global.css'
 import App from './App'
 
 const root = createRoot(document.getElementById('root'))
-
 root.render(
-    // <React.StrictMode>
-        <Router>
-            <Routes>
-                <Route path='/' element={<App />} />
-            </Routes>
-        </Router>
-    // </React.StrictMode>
+  <React.StrictMode>
+    <Router>
+      <Routes>
+        <Route path='/' element={<App />} />
+      </Routes>
+    </Router>
+  </React.StrictMode>
 )
